+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * fix: compatibility with reconstitutor 2.1
 * fix(AtributesPropertyLister): throw exception if there are two properties with
   the same name in the same class hierarchy
+* fix: throw exception if there are two properties with the same name in the
+  same class hierarchy
 
 ## 2.0.0